backend/requirements.txt
django
djangorestframework
pandas
django-cors-headers
reportlab
psycopg2-binary
dj-database-url
python-dotenv
whitenoise
